# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketplace API",
        "description": "Marketplace API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products/get": {
            "get": {
                "tags": [
                    "Товары"
                ],
                "summary": "Получить список товаров",
                "operationId": "c77443495be0269765548567530ea593",
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/products/create": {
            "post": {
                "tags": [
                    "Товары"
                ],
                "summary": "Добавить товар",
                "operationId": "0c967f54238140a6fdca5a3aff16e248",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Название продукта",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Описание продукта",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "description": "Цена продукта",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "category_id": {
                                        "description": "Категория",
                                        "type": "string"
                                    },
                                    "attributes": {
                                        "description": "Атрибуты продукта",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "attribute_name": {
                                                    "description": "Название атрибута",
                                                    "type": "string"
                                                },
                                                "attribute_value": {
                                                    "description": "Значение атрибута",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Продукт успешно добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}/update": {
            "put": {
                "tags": [
                    "Товары"
                ],
                "summary": "Обновить товар",
                "operationId": "ee18e0fc38138ed924908f0e72b09f96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Название товара",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Описание товара",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "description": "Цена товара",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "attributes": {
                                        "description": "Атрибуты товара",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID атрибута",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Название атрибута",
                                                    "type": "string"
                                                },
                                                "value": {
                                                    "description": "Значение атрибута",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Товар успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}/delete": {
            "delete": {
                "tags": [
                    "Товары"
                ],
                "summary": "Удалить товар",
                "operationId": "318bd5ae4047378366e8e94c83a1a28a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Товар успешно удален"
                    },
                    "404": {
                        "description": "Товар не найден"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Схема товара",
                "properties": {
                    "id": {
                        "description": "ID товара",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название товара",
                        "type": "string"
                    },
                    "description": {
                        "description": "Описание товара",
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "description": "Цена товара",
                        "type": "number",
                        "format": "float"
                    },
                    "attributes": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "ID атрибута",
                                    "type": "integer"
                                },
                                "name": {
                                    "description": "Название атрибута",
                                    "type": "string"
                                },
                                "value": {
                                    "description": "Значение атрибута",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}